--- cryptodashboard/src/.env
@@ -1,1 +1,0 @@
-REACT_APP_REST_API = http://localhost:5001/
