+++ cryptodashboard/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App';
+import App from './appStuff/App';
 import CryptoContext from './CryptoContext';
